修改花号基本资料 被用过时 不能删除

+++ src/main/java/com/github/hzw/security/controller/FlowerInfoController.java
@@ -296,10 +296,15 @@ public class FlowerInfoController extends BaseController{
 			String id[] = ids.split(",");
 			for (String string : id) {
 				if(!Common.isEmpty(string)){
-					flowerInfoService.delete(string);
+					FlowerInfo flowerInfo=flowerInfoService.getById(string);
+					if(flowerInfo.getStatus()==1){
+						map.put("flag", "nodelete");
+					}else{
+						map.put("flag", "true");
+						flowerInfoService.delete(string);
+					}
 				}
 			}
-			map.put("flag", "true");
 		} catch (Exception e) {
 			map.put("flag", "false");
 		}
